refactor: Move `check_is_valid_game_path` to `main.rs`
Part of #329
