+++ src-tauri/src/main.rs
@@ -303,6 +303,19 @@ async fn check_is_flightcore_outdated_caller() -> Result<bool, String> {
     check_is_flightcore_outdated().await
 }
 
+/// Checks whether the provided path is a valid Titanfall2 gamepath by checking against a certain set of criteria
+pub fn check_is_valid_game_path(game_install_path: &str) -> Result<(), String> {
+    let path_to_titanfall2_exe = format!("{game_install_path}/Titanfall2.exe");
+    let is_correct_game_path = std::path::Path::new(&path_to_titanfall2_exe).exists();
+    log::info!("Titanfall2.exe exists in path? {}", is_correct_game_path);
+
+    // Exit early if wrong game path
+    if !is_correct_game_path {
+        return Err(format!("Incorrect game path \"{game_install_path}\"")); // Return error cause wrong game path
+    }
+    Ok(())
+}
+
 /// Checks if is valid Titanfall2 install based on certain conditions
 #[tauri::command]
 async fn verify_install_location(game_path: String) -> bool {
